% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fitlmm_f_eq}
\alias{fitlmm_f_eq}
\alias{fitlmm_f_uneq}
\alias{fitlm_2f_eq}
\alias{fitlm_2f_uneq}
\alias{getp_lmm_f}
\alias{getp_lm_2f}
\title{Model fitting helper functions, equal or unequal variances, plus p-value extraction}
\usage{
fitlmm_f_eq(data)

fitlmm_f_uneq(data)

fitlm_2f_eq(data)

fitlm_2f_uneq(data)

getp_lmm_f(model)

getp_lm_2f(model)
}
\arguments{
\item{data}{Dataset to use in analysis}

\item{model}{Fitted lme object for extracting p-value}
}
\description{
Helper functions for fitting models, not currently exported.
}
