% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fitmodel_eq}
\alias{fitmodel_eq}
\alias{fitmodel_uneq}
\alias{getp_lme}
\title{Model fitting helper functions, equal or unequal variances, plus p-value extraction}
\usage{
fitmodel_eq(data)

fitmodel_uneq(data)

getp_lme(model)
}
\arguments{
\item{data}{Dataset to use in analysis}

\item{model}{Fitted lme object for extracting p-value}
}
\description{
Helper functions for fitting models, not currently exported.
}
