% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.simpow}
\alias{print.simpow}
\title{print method for simpow objects}
\usage{
\method{print}{simpow}(x, ...)
}
\arguments{
\item{x}{The output from a power analysis function}

\item{...}{additional arguments, not currently used}
}
\description{
print method for simpow objects
}
