% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpow_lmm_f.R
\name{simpow_lmm_f}
\alias{simpow_lmm_f}
\title{Simulation-based power analysis for a linear mixed model}
\usage{
simpow_lmm_f(
  nsim = 1,
  test = "overall",
  alpha = 0.05,
  ntrt = 2,
  trtmeans,
  trtnames = NULL,
  nrep = 1,
  nblock = 5,
  sd_block,
  sd_resid,
  sd_eq = TRUE,
  keep_data = FALSE,
  keep_models = FALSE
)
}
\arguments{
\item{nsim}{Numeric. The number of simulations to run. Defaults to 1, which is useful if you are testing your approach. Otherwise you'll want to use a large number, such as 1000.}

\item{test}{Character. Can currently take "none" or "overall"; defaults to "overall". Use "none" if you only want simulated datasets. Use "overall" to do power analysis on overall F test of fixed effect.}

\item{alpha}{Numeric. The alpha value you want to test against. Defaults to 0.05.}

\item{ntrt}{Numeric. The number of categories or "treatments" that your categorical fixed effect will have. Defaults to 2. Value must be greater than 1.}

\item{trtmeans}{Numeric vector. The true means for each of your treatment groups.}

\item{trtnames}{Optional character vector. The names for each of your treatment groups. If not provided treatments will be named with letters.}

\item{nrep}{Numeric. Number of replicates within each treatment within each block. Defaults to 1.}

\item{nblock}{Numeric. Total number of blocks, where each treatment is replicated at least one time (based on \code{nrep}) in each block. Defaults to 5. Must be greater than 2.}

\item{sd_block}{Numeric. The true standard deviation based on the among-block variation. The standard deviation is the square root of the variance.}

\item{sd_resid}{Numeric. The true standard deviation of the errors, which you may refer to as the residual standard deviation. The standard deviation is the square root of the variance. If you want to assume equal variance among treatments (the default), provide a single value. If you allow variances to differ among treatments provide a vector that contains a standard deviation for each treatment.}

\item{sd_eq}{Logical. Whether or not to allow the variance of the errors should be constant among treatments. Defaults to \code{TRUE}.}

\item{keep_data}{Logical. Whether you want to keep the simulated datasets that are used in the power analysis. Defaults to \code{FALSE}. It may be useful to keep the datasets if you'd like to do additional exploration of the simulated results.}

\item{keep_models}{Logical. Whether you want to keep the models fit to each simulated datasets for the power analysis. Defaults to \code{FALSE}. It may be useful to keep the models if you'd like to do additional exploration of the models.}
}
\value{
A list containing information on the simulation details, design details and true parameters and, when \code{test = "overall"}, estimated power based on the sample size.
}
\description{
Simulate data and perform a power analysis based on provided study design details and values for parameters. Data must be balanced. Models are fit with \code{nlme::lme()}can contain a single categorical fixed effect and a single random effect. Variances can be allowed to vary among the fixed effect. categories. The categorical fixed effect is referred to as "treatment" and the random effect is called "block".
}
\examples{
# Power analysis based on 100 simulations for the overall test of two treatments.
# There is one replicate of each treatment in each of 10 blocks.
# Note the single residual SD.
simpow_lmm_f(nsim = 100,
             trtmeans = c(1, 25),
             nblock = 10,
             sd_block = 2,
             sd_resid = 4)

# Setting treatment names
simpow_lmm_f(nsim = 100,
             trtmeans = c(1, 25),
             sd_block = 2,
             sd_resid = 4)
}
