% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpow_lmm_f.R
\name{simpow_lmm_f}
\alias{simpow_lmm_f}
\title{Simulation-based power analysis for a linear mixed model}
\usage{
simpow_lmm_f(
  nsim = 1,
  test = "overall",
  alpha = 0.05,
  ntrt = 2,
  trtmeans,
  trtnames = NULL,
  nrep = 1,
  nblock = 5,
  sd_block,
  sd_resid,
  sd_eq = TRUE,
  keep_data = FALSE,
  keep_models = FALSE
)
}
\arguments{
\item{nsim}{Numeric. The number of simulations to run. Defaults to 1, which is useful if you are testing what the datasets look like or how long each analysis takes to run. Otherwise for the full power analysis you'll want to use a large number, such as 1000. Note that the more simulations you do the longer it will take to finish.}

\item{test}{Character. Can currently take "none" or "overall"; defaults to "overall". Use "none" if you only want simulated datasets. Use "overall" to do power analysis on overall F test of fixed effect.}

\item{alpha}{Numeric. The alpha value you want to test against. Defaults to 0.05.}

\item{ntrt}{Numeric. The number of categories or "treatments" that your categorical fixed effect will have. Defaults to 2. Value must be greater than 1.}

\item{trtmeans}{Numeric vector. The true means for each of your treatment groups.}

\item{trtnames}{Optional character vector. The names for each of your treatment groups. If not provided treatments will be named with letters.}

\item{nrep}{Numeric. Number of replicates within each treatment within each block. Defaults to 1.}

\item{nblock}{Numeric. Total number of blocks, where each treatment is replicated at least one time (based on \code{nrep}) in each block. Defaults to 5. Must be greater than 2.}

\item{sd_block}{Numeric. The true standard deviation based on the among-block variation. The standard deviation is the square root of the variance.}

\item{sd_resid}{Numeric. The true standard deviation of the errors, which you may refer to as the residual standard deviation. The standard deviation is the square root of the variance. If you want to assume equal standard deviations among treatments (the default), provide a single value. If you allow standard deviations to differ among treatments provide a vector that contains a standard deviation for each treatment.}

\item{sd_eq}{Logical. Whether or not to allow the variance of the errors should be constant among treatments. Defaults to \code{TRUE}.}

\item{keep_data}{Logical. Whether you want to keep the simulated datasets that are used in the power analysis. Defaults to \code{FALSE}. It may be useful to keep the datasets if you'd like to do additional exploration of the simulated results but this may make output very large.}

\item{keep_models}{Logical. Whether you want to keep the models fit to each simulated datasets for the power analysis. Defaults to \code{FALSE}. It may be useful to keep the models if you'd like to do additional exploration of the models but this may make output very large.}
}
\value{
The printed output contains information on the simulation and power analysis. The object contains a list containing information on the simulation details, design details and true parameters and, when \code{test = "overall"}, estimated power based on the sample size. These can be extracted from the object by name but vary depending on chosen options.\tabular{ll}{
\code{nsim} \tab Number of simulations done. \cr
\code{ntrt} \tab Number of treatment groups. \cr
\code{nblock} \tab Number of blocks in study. \cr
\code{nrep} \tab Number of replications of each treatment group in each block. \cr
\code{truemeans} \tab Named vector of true treatment means used in simulation. \cr
\code{truesd} \tab List containing true block and observation-level (residual) standard deviations used in simulation. \cr
\code{data} \tab If \code{keep_data = TRUE}, the simulated datasets used in the analysis. May be very large. \cr
\code{power} \tab Estimated number of p-values less than alpha, expressed as a percentage. \cr
\code{alpha} \tab Alpha used for power analysis. \cr
\code{p.values} \tab P-values from test for every model. \cr
\code{models} \tab If \code{keep_models = TRUE}, the fitted models for every simulated dataset. May be very large.\cr
}
}
\description{
Simulate data and perform a power analysis based on provided study design details and values for parameters. Data must be balanced. Models are fit with \code{nlme::lme()}can contain a single categorical fixed effect and a single random effect. Variances can be allowed to vary among the fixed effect. categories. The categorical fixed effect is referred to as "treatment" and the random effect is called "block".
}
\examples{
# Power analysis based on 100 simulations for the overall test of two treatments.
# There is one replicate of each treatment in each of 10 blocks.
# Note the single residual SD.
simpow_lmm_f(nsim = 100,
             trtmeans = c(1, 4),
             nblock = 10,
             sd_block = 2,
             sd_resid = 4)


# Allow variances to differ among treatments
simpow_lmm_f(nsim = 100,
             trtmeans = c(1, 4),
             nblock = 10,
             sd_block = 2,
             sd_resid = c(1, 20),
             sd_eq = FALSE)

# Change the number of treatment groups
simpow_lmm_f(nsim = 100,
             ntrt = 3,
             trtmeans = c(1, 3, 4),
             nblock = 10,
             sd_block = 2,
             sd_resid = 4)

 # Return simulated dataset for a single simulation
 # Here don't run power analysis via test = "none"
results = simpow_lmm_f(nsim = 1,
             test = "none",
             trtmeans = c(1, 4),
             nblock = 10,
             sd_block = 2,
             sd_resid = 4,
             keep_data = TRUE)
results$data

\dontrun{
# Setting treatment names to match those in your study
# Seen in results only
results = simpow_lmm_f(nsim = 10,
             trtmeans = c(1, 25),
             trtnames = c("Control", "Treat1")
             sd_block = 2,
             sd_resid = 4)
results
}
}
